do not remove .svn when clear the cache
git-svn-id: https://svn.code.sf.net/p/tikiwiki/code/branches/1.10@12217 b456876b-0849-0410-b77d-98878d47e9d5

+++ lib/cache/cachelib.php
@@ -111,7 +111,7 @@ class Cachelib {
 	if (!$path or !is_dir($path)) return 0;
 	if ($dir = opendir($path)) {
 		while (false !== ($file = readdir($dir))) {
-			if (substr($file,0,1) == "." or $file == 'CVS' or $file == "index.php" or $file == "README") continue;
+			if (substr($file,0,1) == "." or $file == 'CVS' or $file == "index.php" or $file == "README" or $file == '.svn') continue;
 			if (is_dir($path."/".$file)) {
 				$this->erase_dir_content($path."/".$file);
 				rmdir($path."/".$file);
